rename file and allow passing a boolet from outside, also fixed help text
